+++ src/components/form/go-switch/demo/go-switch.html
