+++ src/components/form/go-switch/go-switch.scss
