+++ src/components/form/go-switch/readme.md
