+++ src/components/form/go-switch/test/go-switch.spec.tsx
