+++ src/index.html
@@ -17,7 +17,7 @@
     <!-- keep this sandbox empty before pushing changes -->
 
     <!-- #region demo start -->
-
+    <go-switch checked name="switch" blah="blah" label="Test switch" class="test" style="display: block"></go-switch>
     <!-- #endregion demo finish -->
 
     <aside>
