+++ src/interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './variants';
 export * from './nav-item';
 export * from './icon';
+export * from './form';
